Fix: Um Veterinário só deve conseguir deletar Exames se ele pertencer a Consulta

+++ src/main/java/app/com/_paws/services/ExamService.java
@@ -1,9 +1,14 @@
 package app.com._paws.services;
 
+import app.com._paws.domain.models.Exam;
+import app.com._paws.domain.models.Veterinarian;
 import app.com._paws.domain.repositories.ExamRepository;
+import app.com._paws.exceptions.NotFoundException;
+import app.com._paws.exceptions.PermissionException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
+import java.util.List;
 import java.util.UUID;
 
 @Service
@@ -11,8 +16,26 @@ import java.util.UUID;
 public class ExamService {
 
     private final ExamRepository examRepository;
+    private final AuthService authService;
 
     public void deleteById(UUID id) {
-        this.examRepository.deleteById(id);
+
+        Exam exam = this.examRepository.findById(id)
+                .orElseThrow(() -> new NotFoundException("Exame não encontrado!"));
+
+        UUID vetUUID = this.authService.obtainAuthenticatedUserUUID();
+        List<Veterinarian> veterinarians = exam.getAppointment().getVeterinarians();
+        boolean isVetAuthorized = veterinarians
+                .stream()
+                .anyMatch(
+                        veterinarian -> veterinarian.getId().equals(vetUUID)
+                );
+
+        if (isVetAuthorized) {
+            this.examRepository.deleteById(id);
+        }
+        else {
+            throw new PermissionException("Você não tem autorização para deletar esse Exame!");
+        }
     }
 }
